+++ src/main/java/gift/controller/api/MemberController.java
@@ -1,6 +1,9 @@
 package gift.controller;
 
-import gift.dto.*;
+import gift.dto.MemberRequestDto;
+import gift.dto.MemberResponseDto;
+import gift.dto.MemberRoleRequestDto;
+import gift.dto.TokenResponseDto;
 import gift.service.MemberService;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
